* code refactor
