+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:mobile_banking_system/src/widgets/home_screen.dart';
+import 'package:mobile_banking_system/src/routes/home_screen.dart';
 
 main() {
   runApp(MyApp());
