+++ lib/src/routes/home_screen.dart
@@ -1,9 +1,9 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:mobile_banking_system/src/widgets/app_drawer.dart';
-import 'package:mobile_banking_system/src/widgets/balance_page.dart';
-import 'package:mobile_banking_system/src/widgets/transfer_funds_page.dart';
+import 'package:mobile_banking_system/src/app_drawer/app_drawer.dart';
+import 'package:mobile_banking_system/src/routes/balance_page.dart';
+import 'package:mobile_banking_system/src/routes/transfer_funds_page.dart';
 
 class HomeScreen extends StatefulWidget {
   final GoogleSignIn googleSignIn;
