--- lib/src/widgets/statement_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class StatementPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("statement page"),
-    );
-  }
-}
