--- lib/src/widgets/support_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SupportPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("support page"),
-    );
-  }
-}
