+++ pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
   cupertino_icons: ^0.1.2
   firebase_auth: ^0.7.0
   google_sign_in: ^3.2.4
+  url_launcher: ^5.0.2
